@@ -1,7 +1,7 @@
 {
   "name": "react-native-share",
   "description": "Social share, sending simple data to other apps.",
-  "version": "8.0.0",
+  "version": "10.0.2",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/react-native-community/react-native-share.git"
@@ -12,45 +12,56 @@
   "main": "lib/commonjs/index.js",
   "react-native": "src/index.tsx",
   "module": "lib/module/index.js",
-  "types": "lib/typescript/index.d.ts",
+  "types": "lib/typescript/src/index.d.ts",
   "files": [
     "src",
     "lib",
-    "!**/__tests__",
-    "!**/__fixtures__",
-    "!**/__mocks__",
     "android",
     "ios",
     "windows",
-    "RNShare.podspec"
+    "codegenSpec",
+    "RNShare.podspec",
+    "!**/__tests__",
+    "!**/__fixtures__",
+    "!**/__mocks__"
   ],
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@babel/runtime": "^7.12.5",
-    "@commitlint/cli": "11.0.0",
-    "@commitlint/config-conventional": "11.0.0",
-    "@react-native-community/eslint-config": "^3.0.0",
-    "@semantic-release/commit-analyzer": "^8.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@semantic-release/github": "^7.1.1",
-    "@semantic-release/npm": "^7.0.6",
-    "@semantic-release/release-notes-generator": "^9.0.1",
-    "@types/react": "16.14.5",
-    "@types/react-native": "0.63.2",
-    "eslint": "^7.24.0",
-    "eslint-config-satya164": "3.1.10",
-    "husky": "4.3.0",
-    "lint-staged": "10.3.0",
-    "metro-config": "^0.66.2",
-    "metro-react-native-babel-preset": "^0.66.2",
-    "pre-commit": "1.2.2",
-    "prettier": "^2.2.1",
-    "react": "17.0.2",
-    "react-native": "0.66.4",
-    "react-native-builder-bob": "^0.18.1",
-    "react-native-windows": "0.66.4",
-    "semantic-release": "^19.0.3",
-    "typescript": "4.2.4"
+    "@babel/core": "^7.23.3",
+    "@babel/preset-env": "^7.23.3",
+    "@babel/runtime": "^7.23.2",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
+    "@react-native/eslint-config": "^0.74.0",
+    "@react-native/metro-config": "^0.73.2",
+    "@semantic-release/commit-analyzer": "^11.1.0",
+    "@semantic-release/git": "^10.0.1",
+    "@semantic-release/github": "^9.2.1",
+    "@semantic-release/npm": "^11.0.1",
+    "@semantic-release/release-notes-generator": "^12.1.0",
+    "@tsconfig/react-native": "^3.0.2",
+    "@types/react": "^18.2.37",
+    "@types/react-native": "^0.72.6",
+    "@types/react-test-renderer": "^18.0.6",
+    "babel-jest": "^29.7.0",
+    "eslint": "^8.53.0",
+    "eslint-config-satya164": "^3.2.0",
+    "husky": "^4.3.0",
+    "jest": "^29.7.0",
+    "lint-staged": "^15.0.2",
+    "metro-config": "^0.80.0",
+    "metro-react-native-babel-preset": "^0.77.0",
+    "pre-commit": "^1.2.2",
+    "prettier": "^3.0.3",
+    "react": "^18.2.0",
+    "react-native": "0.72.6",
+    "react-native-builder-bob": "^0.23.1",
+    "react-native-windows": "^0.72.19",
+    "react-test-renderer": "^18.2.0",
+    "semantic-release": "^22.0.7",
+    "typescript": "^5.2.2"
+  },
+  "resolutions": {
+    "eslint-plugin-prettier": "5.0.1"
   },
   "keywords": [
     "react-native",
@@ -77,8 +88,8 @@
   "scripts": {
     "start": "react-native start",
     "start:android": "react-native run-android",
-    "start:ios": "react-native run-ios --project-path example/ios",
-    "lint": "eslint \"src/**/*.{js,ts,tsx}\" --max-warnings=0",
+    "start:ios": "react-native run-ios",
+    "lint": "eslint \"src/**/*.{js,ts,tsx}\" \"codegenSpec/**/*.{js,ts,tsx}\" --max-warnings=0",
     "typescript": "tsc --noEmit",
     "validate": "yarn lint && yarn typescript",
     "ci:publish": "npx semantic-release",
@@ -125,5 +136,16 @@
       "module",
       "typescript"
     ]
+  },
+  "codegenConfig": {
+    "name": "RNShareSpec",
+    "type": "modules",
+    "jsSrcsDir": "./codegenSpec",
+    "android": {
+      "javaPackageName": "cl.json"
+    }
+  },
+  "engines": {
+    "node": ">=16"
   }
 }
