@@ -18,8 +18,8 @@
 #import <React/RCTUtils.h>
 @interface InstagramShare : NSObject <RCTBridgeModule>
 
-- (void) shareSingle:(NSDictionary *)options failureCallback:(RCTResponseErrorBlock)failureCallback successCallback:(RCTResponseSenderBlock)successCallback;
+- (void) shareSingle:(NSDictionary *)options reject:(RCTPromiseRejectBlock)reject resolve:(RCTPromiseResolveBlock)resolve;
 - (void)shareSingleImage:(NSDictionary *)options
-         failureCallback:(RCTResponseErrorBlock)failureCallback
-         successCallback:(RCTResponseSenderBlock)successCallback;
+         reject:(RCTPromiseRejectBlock)reject
+         resolve:(RCTPromiseResolveBlock)resolve;
 @end
