@@ -1,20 +1,22 @@
-import { NativeModules, Platform } from 'react-native';
+import NativeRNShare from '../codegenSpec/NativeRNShare';
 
 import Overlay from './components/Overlay';
 import Sheet from './components/Sheet';
 import Button from './components/Button';
 import ShareSheet from './components/ShareSheet';
-import requireAndAskPermissions from './helpers/requireAndAskPermissions';
+import checkPermissions from './helpers/checkPermissions';
 import {
   Social,
   IsPackageInstalledResult,
   ActivityType,
-  ShareOpenResult,
   ShareAsset,
+  ShareOpenResult,
   ShareOptions,
   ShareSingleOptions,
   ShareSingleResult,
 } from './types';
+import { isAndroid, isIOS } from './helpers/platform';
+import { normalizeShareOpenOptions, normalizeSingleShareOptions } from './helpers/options';
 
 const RNShare = {
   Button,
@@ -23,124 +25,80 @@ const RNShare = {
   Sheet,
 
   Social: {
-    FACEBOOK: NativeModules.RNShare.FACEBOOK || Social.Facebook,
-    FACEBOOK_STORIES: NativeModules.RNShare.FACEBOOKSTORIES || Social.FacebookStories,
-    PAGESMANAGER: NativeModules.RNShare.PAGESMANAGER || Social.Pagesmanager,
-    TWITTER: NativeModules.RNShare.TWITTER || Social.Twitter,
-    WHATSAPP: NativeModules.RNShare.WHATSAPP || Social.Whatsapp,
-    WHATSAPPBUSINESS: NativeModules.RNShare.WHATSAPPBUSINESS || Social.Whatsappbusiness,
-    INSTAGRAM: NativeModules.RNShare.INSTAGRAM || Social.Instagram,
-    INSTAGRAM_STORIES: NativeModules.RNShare.INSTAGRAMSTORIES || Social.InstagramStories,
-    GOOGLEPLUS: NativeModules.RNShare.GOOGLEPLUS || Social.Googleplus,
-    EMAIL: NativeModules.RNShare.EMAIL || Social.Email,
-    PINTEREST: NativeModules.RNShare.PINTEREST || Social.Pinterest,
-    LINKEDIN: NativeModules.RNShare.LINKEDIN || Social.Linkedin,
-    SMS: NativeModules.RNShare.SMS || Social.Sms,
-    TELEGRAM: NativeModules.RNShare.TELEGRAM || Social.Telegram,
-    MESSENGER: NativeModules.RNShare.MESSENGER || Social.Messenger,
-    SNAPCHAT: NativeModules.RNShare.SNAPCHAT || Social.Snapchat,
-    VIBER: NativeModules.RNShare.VIBER || Social.Viber,
+    FACEBOOK: NativeRNShare.getConstants().FACEBOOK || Social.Facebook,
+    FACEBOOK_STORIES: NativeRNShare.getConstants().FACEBOOKSTORIES || Social.FacebookStories,
+    PAGESMANAGER: NativeRNShare.getConstants().PAGESMANAGER || Social.Pagesmanager,
+    TWITTER: NativeRNShare.getConstants().TWITTER || Social.Twitter,
+    WHATSAPP: NativeRNShare.getConstants().WHATSAPP || Social.Whatsapp,
+    WHATSAPPBUSINESS: NativeRNShare.getConstants().WHATSAPPBUSINESS || Social.Whatsappbusiness,
+    INSTAGRAM: NativeRNShare.getConstants().INSTAGRAM || Social.Instagram,
+    INSTAGRAM_STORIES: NativeRNShare.getConstants().INSTAGRAMSTORIES || Social.InstagramStories,
+    GOOGLEPLUS: NativeRNShare.getConstants().GOOGLEPLUS || Social.Googleplus,
+    EMAIL: NativeRNShare.getConstants().EMAIL || Social.Email,
+    PINTEREST: NativeRNShare.getConstants().PINTEREST || Social.Pinterest,
+    LINKEDIN: NativeRNShare.getConstants().LINKEDIN || Social.Linkedin,
+    SMS: NativeRNShare.getConstants().SMS || Social.Sms,
+    TELEGRAM: NativeRNShare.getConstants().TELEGRAM || Social.Telegram,
+    MESSENGER: NativeRNShare.getConstants().MESSENGER || Social.Messenger,
+    SNAPCHAT: NativeRNShare.getConstants().SNAPCHAT || Social.Snapchat,
+    VIBER: NativeRNShare.getConstants().VIBER || Social.Viber,
   },
 
-  open(options: ShareOptions): Promise<ShareOpenResult | never> {
-    return new Promise((resolve, reject) => {
-      requireAndAskPermissions(options)
-        .then(() => {
-          if (Platform.OS === 'ios' && options.url && !options.urls) {
-            // Backward compatibility with { Share } from react-native
-            const url = options.url;
-            delete options.url;
+  async open(options: ShareOptions) {
+    await checkPermissions(options);
 
-            options.urls = [url];
+    options = normalizeShareOpenOptions(options);
 
-            if (options.filename && !options.filenames) {
-              options.filenames = [options.filename];
-              delete options.filename;
-            }
-          }
+    const result: ShareOpenResult = await NativeRNShare.open(options);
 
-          NativeModules.RNShare.open(
-            options,
-            (error) => {
-              return reject({ error });
-            },
-            (success, message) => {
-              if (success) {
-                return resolve({
-                  success,
-                  message,
-                });
-              } else if (options.failOnCancel === false) {
-                return resolve({
-                  dismissedAction: true,
-                  success,
-                  message,
-                });
-              } else {
-                reject(new Error('User did not share'));
-              }
-            },
-          );
-        })
-        .catch((e: unknown) => reject(e));
-    });
-  },
+    if (!result.success) {
+      if (options.failOnCancel) {
+        throw new Error('User did not share');
+      }
 
-  shareSingle(options: ShareSingleOptions): Promise<ShareSingleResult | never> {
-    if (Platform.OS === 'ios' || Platform.OS === 'android') {
-      return new Promise((resolve, reject) => {
-        requireAndAskPermissions(options)
-          .then(() => {
-            if (options.url) {
-              options.urls = [options.url];
-            }
-
-            if (options.social === RNShare.Social.INSTAGRAM_STORIES && !options.appId) {
-              return reject({
-                success: false,
-                message: 'Instagram Story share requires an appId based on Meta policy.',
-              });
-            }
-
-            NativeModules.RNShare.shareSingle(
-              options,
-              (error) => {
-                return reject({ error });
-              },
-              (success, message) => {
-                return resolve({
-                  success: Boolean(success),
-                  message,
-                });
-              },
-            );
-          })
-          .catch((e: unknown) => reject(e));
-      });
-    } else {
-      throw new Error('Not implemented');
+      const dismissedResult: ShareOpenResult = {
+        dismissedAction: true,
+        success: result.success,
+        message: result.message,
+      };
+      return dismissedResult;
     }
+
+    return result;
   },
 
-  isPackageInstalled(packageName: string): Promise<IsPackageInstalledResult | never> {
-    if (Platform.OS === 'android') {
-      return new Promise((resolve, reject) => {
-        NativeModules.RNShare.isPackageInstalled(
-          packageName,
-          (error) => {
-            return reject({ error });
-          },
-          (isInstalled) => {
-            return resolve({
-              isInstalled,
-              message: 'Package is Installed',
-            });
-          },
-        );
-      });
-    } else {
-      throw new Error('Not implemented');
+  async shareSingle(options: ShareSingleOptions) {
+    if (!isAndroid() && !isIOS()) throw new Error('Not implemented');
+
+    if (options.social === RNShare.Social.INSTAGRAM_STORIES && !options.appId) {
+      throw new Error('To share to Instagram Stories you need to provide appId');
     }
+
+    await checkPermissions(options);
+
+    options = normalizeSingleShareOptions(options);
+
+    const { success, message } = await NativeRNShare.shareSingle(options);
+
+    const result: ShareSingleResult = {
+      success: Boolean(success),
+      message,
+    };
+
+    return result;
+  },
+
+  async isPackageInstalled(packageName: string) {
+    if (!isAndroid()) throw new Error('Not implemented');
+
+    const isInstalled = await NativeRNShare.isPackageInstalled(packageName);
+
+    const result: IsPackageInstalledResult = {
+      isInstalled,
+      message: 'Package is Installed',
+    };
+
+    return result;
   },
 } as const;
 
